Moved validation functions to its own file, and fixed regex
