+++ sdk/v2/core/projects_test.go
@@ -254,45 +254,3 @@ func TestProjectsClientDelete(t *testing.T) {
 	err := client.Delete(context.Background(), testProjectID)
 	require.NoError(t, err)
 }
-
-func TestValidateProjectID(t *testing.T) {
-	testCases := []struct {
-		name     string
-		id       string
-		expected bool
-	}{
-		{
-			name:     "valid project id",
-			id:       "abc",
-			expected: true,
-		},
-		{
-			name:     "id is too short",
-			id:       "aa",
-			expected: false,
-		},
-		{
-			name: "id is too long by 1 character",
-			id: "aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa" +
-				"aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
-			expected: false,
-		},
-		{
-			name:     "id starts with a non-letter character",
-			id:       "1aa",
-			expected: false,
-		},
-		{
-			name:     "id contains a capital letter",
-			id:       "Hello",
-			expected: false,
-		},
-	}
-
-	for _, testCase := range testCases {
-		t.Run(testCase.id, func(t *testing.T) {
-			err := ValidateProjectID(testCase.id)
-			require.Equal(t, testCase.expected, err == nil)
-		})
-	}
-}
