+++ v2/cli/init_command.go
@@ -5,7 +5,6 @@ import (
 	"io/ioutil"
 	"os"
 	"path"
-	"regexp"
 	"strings"
 
 	"github.com/AlecAivazis/survey/v2"
@@ -63,9 +62,11 @@ func initialize(c *cli.Context) error {
 		return err
 	}
 
-	err = validateGitCloneURL(fields.GitCloneURL)
-	if err != nil {
-		return err
+	if fields.GitCloneURL != "" {
+		err = core.ValidateGitCloneURL(fields.GitCloneURL)
+		if err != nil {
+			return err
+		}
 	}
 
 	// Check if input language is valid
@@ -315,18 +316,4 @@ func checkFileOverwrite(
 		return response, true, nil
 	}
 	return true, false, nil
-}
-
-func validateGitCloneURL(url string) error {
-	gitCloneURLRegex :=
-		regexp.MustCompile(
-			`((git|ssh|http(s)?)|(git@[\w\.]+))` +
-				`(:(//)?)([\w\.@\:/\-~]+)(\.git)(/)?`)
-	urlMatch := gitCloneURLRegex.MatchString(url)
-	if !urlMatch {
-		return fmt.Errorf("invalid value %q for git clone URL"+
-			" (must start with http://, https://, or git@"+
-			" and must end with .git)", url)
-	}
-	return nil
 }
